orderdetail-addition

+++ server/src/main/java/com/server/app/dto/OrderDto.java
@@ -20,7 +20,7 @@ public class OrderDto {
     private LocalDate orderDate;
     private LocalDate requiredDate;
     private LocalDate shippedDate;
-    private Float freight;
+    private Double freight;
     private String shipName;
     private String shipAddress;
     private String shipCity;
