+++ server/src/main/java/com/server/app/dto/request/OrderUpdateRequest.java
@@ -21,7 +21,7 @@ public class OrderUpdateRequest {
     private LocalDate orderDate;
     private LocalDate requiredDate;
     private LocalDate shippedDate;
-    private Float freight;
+    private Double freight;
     private String shipName;
     private String shipAddress;
     private String shipCity;
